+++ lib/schema_to_hash/column.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module SchemaToHash
+  class Column
+    attr_reader :name, :type
+
+    def initialize(name:, type:)
+      @name = name
+      @type = type
+    end
+  end
+end
