+++ lib/schema_to_hash/type.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module SchemaToHash
+  TYPE = [
+    'bigint',
+    'binary',
+    'boolean',
+    'date',
+    'datetime',
+    'decimal',
+    'float',
+    'integer',
+    'string',
+    'text'
+  ]
+end
