--- lib/schema_to_json/column.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class Column
-  attr_reader :name, :type
-
-  def initialize(name:, type:)
-    @name = name
-    @type = type
-  end
-end
