--- lib/schema_to_json/table_list.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-class TableList
-  attr_reader :tables
-
-  def initialize(tables = [])
-    @tables = tables
-  end
-
-  def add(table)
-    @tables << table
-  end
-end
