--- lib/schema_to_json/type.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-TYPE = [
-  'bigint',
-  'binary',
-  'boolean',
-  'date',
-  'datetime',
-  'decimal',
-  'float',
-  'integer',
-  'string',
-  'text'
-]
